- Made it so you can now add books. Quick commit message, but essentially just made it so you can now add books. Also the addBookModule now takes authors as an inputs.

+++ public_html/listDataModule.php
@@ -100,7 +100,7 @@
                 echo "<td>" . $pdo->quote($publisherName) . "</td>";
                 echo "<td>" . $pdo->quote($formatName) . "</td>";
                 echo "<td>" . sanitize($pdo, $book['year']) . "</td>";
-                $read = sanitize($pdo, $book['haveRead']) == 1 ? "Yes" : "No";
+                $read = (int)sanitize($pdo, $book['haveRead']) == 1 ? "Yes" : "No";
                 echo "<td>" . $read . "</td>";
 
                 // Adds a delete button to each row
